renames app from reactnativestareterkit to learnacademy.mobile in android
