+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.reactnativestarterkit",
+    package = "com.learnacademy.mobile",
 )
 
 android_resource(
     name = "res",
-    package = "com.reactnativestarterkit",
+    package = "com.learnacademy.mobile",
     res = "src/main/res",
 )
